requirements.txt
flask
pymongo
werkzeug
mongomock
yfinance
python-dotenv
email-validator
google-generativeai
